requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.3
uwsgi==2.0.20
drf-yasg==1.21.4
django-split-settings==1.1.0
django-filter==22.1
